Moved context into a separate file
