+++ crates/modelardb_server/src/metadata.rs
@@ -45,14 +45,15 @@ use sqlx::{Executor, Result, Row, Sqlite, SqlitePool};
 use tracing::{error, info, warn};
 use uuid::Uuid;
 
+use crate::context::Context;
 use crate::query::ModelTable;
 use crate::storage::COMPRESSED_DATA_FOLDER;
-use crate::Context;
 
 /// Name used for the file containing the SQLite database storing the metadata.
 pub const METADATA_DATABASE_NAME: &str = "metadata.sqlite3";
 
 /// Metadata about a file tracked by [`MetadataManager`] which contains compressed segments.
+#[allow(dead_code)]
 #[derive(Debug, Clone)]
 pub struct CompressedFile {
     /// Name of the file.
@@ -67,6 +68,7 @@ pub struct CompressedFile {
     max_value: Value,
 }
 
+#[allow(dead_code)]
 impl CompressedFile {
     pub fn new(
         name: Uuid,
@@ -327,6 +329,7 @@ impl MetadataManager {
     /// Compute the 64-bit univariate ids of the univariate time series to retrieve from the storage
     /// engine using the field columns, tag names, and tag values in the query. Returns a [`Error`]
     /// if the necessary data cannot be retrieved from the metadata database.
+    #[allow(dead_code)]
     pub async fn compute_univariate_ids_using_fields_and_tags(
         &self,
         table_name: &str,
@@ -432,6 +435,7 @@ impl MetadataManager {
     /// * The max value is smaller than the min value in `compressed_file`.
     /// * The metadata database could not be modified.
     /// * A model table with `model_table_name` does not exist.
+    #[allow(dead_code)]
     pub async fn save_compressed_file(
         &self,
         model_table_name: &str,
@@ -464,6 +468,7 @@ impl MetadataManager {
     /// * Less than the number of files in `compressed_files_to_delete` was deleted.
     /// * The metadata database could not be modified.
     /// * A model table with `model_table_name` does not exist.
+    #[allow(dead_code)]
     pub async fn replace_compressed_files(
         &self,
         model_table_name: &str,
@@ -591,6 +596,7 @@ impl MetadataManager {
     /// * The max value is smaller than the min value.
     /// * The metadata database could not be accessed.
     /// * A model table with `model_table_name` does not exist.
+    #[allow(dead_code)]
     pub async fn compressed_files(
         &self,
         model_table_name: &str,
